# docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
  handlers.RejectRequest:
    properties:
      reason:
        example: Muy lejos
        type: string
    type: object
  httpx.JSendError:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: Failed to connect to database
        type: string
      status:
        example: error
        type: string
    type: object
  httpx.JSendFail:
    properties:
      data:
        additionalProperties: {}
        type: object
      status:
        example: fail
        type: string
    type: object
...
  httpx.JSendSuccess:
    properties:
      data:
        type: object
      status:
        example: success
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Delivery marketplace API for customers, merchants, and drivers.
  title: TacoShare Delivery API
  version: "1.0"
paths:
  /assignments/{order_id}/accept:
    post:
      consumes:
      - application/json
      description: Accept a pending order assignment
      parameters:
      - description: Order ID (UUID)
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Assignment accepted
          schema:
            $ref: '#/definitions/httpx.JSendSuccess'
        "400":
          description: Invalid order ID
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "404":
          description: Assignment not found
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/httpx.JSendError'
      security:
      - BearerAuth: []
      summary: Accept assignment
      tags:
      - assignments
  /assignments/{order_id}/reject:
    post:
      consumes:
      - application/json
      description: Reject a pending order assignment
      parameters:
      - description: Order ID (UUID)
        in: path
        name: order_id
        required: true
        type: string
      - description: Reject reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RejectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Assignment rejected
          schema:
            $ref: '#/definitions/httpx.JSendSuccess'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/httpx.JSendError'
      security:
      - BearerAuth: []
      summary: Reject assignment
      tags:
      - assignments
  /assignments/pending:
    get:
      consumes:
      - application/json
      description: Get all pending assignments for the authenticated driver
      produces:
      - application/json
      responses:
        "200":
          description: Pending assignments retrieved
          schema:
            $ref: '#/definitions/httpx.JSendSuccess'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/httpx.JSendError'
      security:
      - BearerAuth: []
      summary: Get pending assignments for driver
      tags:
      - assignments
...
produces:
- application/json
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"